After I discovered what benefit Guice and JCommander would have on our project, I decided to use both of them.
In the short run, this was a pretty big setback as I had to rewrite almost everything I had started. But in the long run, we now have the ability to unit test the CLI project with ease. No more worrying about lots of exception handling and trying to mock out your mom and the kitchen sink. Just some plain owld guice goodness comming at ya.

Also the commands that we write from now on will be contained to one file, which is much better than the route we were going down before. Here's to configuration over composition!