General Hygiene and Cleanup
Removed some unused imports across multiple files.
Added the ability for each command to print out its own help text using a "-help" flag
Moved some of the configuration code into the CommandLineParsingService.java file so that we could
	get around some of the limitations of jCommander.
Added a Default provider configuration interface so that in the future we can do more with default 
	configurations with properties files.
Modified a few tests that were failing when compiled on our linux systems.