Started Configuration for JSDoc, will need to wait for Kaizer to get this working with Hudson.
Added Filtering to the ReADME, LICENSE and NOTICE files even though there is nothing inside them yet. THis will help us maintain these files when we make updates in the future.

Added Eclipse Standard Formatting to all source files.

Implemented the run tests command. This is only part of the work. still need to implement the cleanup of the Automation Results.plist file when the tests are done running.