modified the pom to include links back to the appropriate sections of our project site.
Added template file validation to CommandRunTests.java Also added validation for required fields within the execution code. Currently, JCommander does not support this so we had to work around it.

Tweaked One or two fields in the IOSAutomationReporting Service.java class Also added some logging facilities so that we can better diagnose things in the future.

Modified the descriptions of the errors thrown by the JavaScriptFileValidator.

Modified the TestRunner.js file to catch exceptions during setUpTestSet() and tearDownTestSet(). Also added the ability to skip parts of the testing life cycle. This will help when reporting on tests that are captured in comments and run manually.

Added IMAT settings to support the life cycle changes

Modified the StemView.js class to be a part of the IMAT namespace.